Update api to use common validates
