+++ src/app/modules/create-facility/services/create-facility-api.service.ts
@@ -5,10 +5,6 @@ import { CreateFacilityDataService } from './create-facility-data.service';
 import { CommonImage } from 'moh-common-lib';
 import { flatMap, catchError } from 'rxjs/operators';
 import { BCPDocumentTypes } from '../../core-bcp/models/documentTypes';
-import {
-  PractitionerValidationPartial,
-  ValidatePractitionerRequest,
-  FacilityValidationPartial } from '../../core-bcp/models/base-api.model';
 import { HttpClient } from '@angular/common/http';
 
 @Injectable({
@@ -20,37 +16,6 @@ export class CreateFacilityApiService extends BCPApiService {
               protected logger: SplunkLoggerService,
               protected dataService: CreateFacilityDataService) {
     super(http, logger, dataService);
-  }
-
-  // TODO:  If flag can be added to indicate to look for MD, this will need to be
-  //        moved to the BCPApiService
-  validatePractitioner(practitioner: PractitionerValidationPartial, applicationUUID) {
-    const payload: ValidatePractitionerRequest = {
-      practitioner,
-      requestUUID: this.generateUUID(),
-      applicationUUID
-    };
-    this.dataService.jsonApplicantValidation.request = payload;
-
-    const url = `${this.baseUrl}/validatePractitioner`;
-
-    return this.post(url, payload);
-  }
-
-  // TODO:  If can check whether field exists or not, this will need to be
-  //        moved to the BCPApiService
-  validateFacility(facility: FacilityValidationPartial, applicationUUID: string) {
-    const payload = {
-      facility,
-      requestUUID: this.generateUUID(),
-      applicationUUID
-    };
-
-    this.dataService.jsonFacilityValidation.request = payload;
-
-    const url = `${this.baseUrl}/validateFacility`;
-
-    return this.post(url, payload);
   }
 
   /**
