+++ src/app/services/bcp-api.service.ts
@@ -5,6 +5,7 @@ import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http
 import { SplunkLoggerService } from './splunk-logger.service';
 import { BCPDocumentTypes } from '../modules/core-bcp/models/documentTypes';
 import { BaseDataService } from './base-data.service';
+import { PractitionerValidationPartial, ValidatePractitionerRequest, FacilityValidationPartial } from '../modules/core-bcp/models/base-api.model';
 
 // TODO: Type Requests
 // TODO: Type responses
@@ -51,6 +52,34 @@ export class BCPApiService extends AbstractHttpService {
     url += `?attachmentdocumenttype=SIGNATURE&programArea=CLAIMS&contentType=1`;
 
     return this.uploadAttachment(url, attachment);
+  }
+
+  // Validate Practitioner/MD - If validate MD, doctor flag must be true, otherwise false
+  validatePractitioner(practitioner: PractitionerValidationPartial, applicationUUID) {
+    const payload: ValidatePractitionerRequest = {
+      practitioner,
+      requestUUID: this.generateUUID(),
+      applicationUUID
+    };
+    this.dataService.jsonApplicantValidation.request = payload;
+
+    const url = `${this.baseUrl}/validatePractitioner`;
+
+    return this.post(url, payload);
+  }
+
+  validateFacility(facility: FacilityValidationPartial, applicationUUID: string) {
+    const payload = {
+      facility,
+      requestUUID: this.generateUUID(),
+      applicationUUID
+    };
+
+    this.dataService.jsonFacilityValidation.request = payload;
+
+    const url = `${this.baseUrl}/validateFacility`;
+
+    return this.post(url, payload);
   }
 
 }
