+++ src/version.GENERATED.ts
@@ -1,9 +1,9 @@
 // DO NOT DELETE OR APP WILL FAIL TO COMPILE! Generated from version.js
-export const gitCommit = 'afc604d';
-export const buildTime = '1/7/2020 at 12:42:03 PM';
+export const gitCommit = 'ad8a5bf';
+export const buildTime = '1/7/2020 at 2:59:18 PM';
 /** App version retrieved from package.json. */
 export const projectVersion = '0.10.1';
 /** Human readable message  */
-export const message = 'Canonical Version: afc604d - 1/7/2020 at 12:42:03 PM (v0.10.1) \nhttps://github.com/bcgov/moh-bcp/commits/afc604d';
+export const message = 'Canonical Version: ad8a5bf - 1/7/2020 at 2:59:18 PM (v0.10.1) \nhttps://github.com/bcgov/moh-bcp/commits/ad8a5bf';
 /** If true, other values should be present. */
 export const success = true;
